pyproject.toml
[project]
name = "xiaohongshu-product-poster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pillow>=12.1.1",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "rich>=14.3.3",
    "typer>=0.24.1",
]

[project.scripts]
xhs-poster = "xhs_poster.cli:main"
